Shoot controller starts only after all Project resources have been created
Now that we have Project resources we should wait for the Project controller to create the Project resources
for all namespaces before starting the Shoot controller. This is because the Shoot controller wants to read
the Project object responsible for a respective Shoot object and, if it starts simultaneously to the Project
controller, some of the needed objects are not yet created.
This is only needed for backwards-compatibility and can be removed in a future version.