Grafana dashboards rework
All existing dashboards are re-worked to function properly and new many of the
graphs are consolidated.

Following dashboards are now available:

- Alertmanager Alerts with small collection of graphs for current and previous
alerts.
- Kubernetes Cluster Health presents general availabity of some of the core
components with links to Alertmanager Alerts dashboard.
- Kubernetes Control Plane Status focuses on the control plane components - API
Server, Scheduler and Controller manager with API Server latency and fail rate.
- Kubernetes Deployments with all deployments in the Shoot's namespace in the
Seed cluster and core components in the Shoot's kube-system namespace. It's
Possible to drill down to individual Pods in the Kubernetes Pods dashboard
- Kubernetes Pods focuses on individual Pods in either Seed or Shoot cluster.
- Prometheus dashboard with various Prometheus metrics.
- Shoot Nodes for individual Node metrics (capacity, allocations), container and
systemd services.