Add support to AWS cloud-specific re-advertiser
Currently the re-advertiser updates the deployment of the api-server with the new advertise address seen from the AWS ELB, this causes the API server to restart frequently.

This pull request, updates this functionality by allowing the re-advertiser to edit the cluster endpoint directly assuming the api-server --endpoint-reconciler-type is set to none as the aws-lb-readvertiser will be reconciling the endpoints itself. This way, less api-restarts will be achieved.

Furthermore, the `advertise-address` is now set per cloud-provider as in case of AWS the `aws-lb-readvertiser` sets the value of the endpoints /
create them itself.